+++ xknx/secure/data_secure.py
@@ -212,6 +212,7 @@ class DataSecure:
                 scf=s_apdu.scf,
                 address_fields_raw=_address_fields_raw,
                 address_type=cemi_data.address_type,
+                frame_format=cemi_data.flags.frame_format,
                 tpci=cemi_data.tpci,
             )
         decrypted_payload = APCI.from_knx(plain_apdu_raw)
@@ -260,6 +261,7 @@ class DataSecure:
             address_fields_raw=cemi_data.src_addr.to_knx()
             + cemi_data.dst_addr.to_knx(),
             address_type=cemi_data.address_type,
+            frame_format=cemi_data.flags.frame_format,
             tpci=cemi_data.tpci,
         )
         secure_cemi_data = copy(cemi_data)